Rename output format test helper
