+++ test/unit/output_formats/output_format_test_helper.rb
