+++ test/unit/output_formats/test_naked_srcset.rb
@@ -1,4 +1,4 @@
-require_relative './test_helper_output'
+require_relative './output_format_test_helper'
 
 class TestNakedSrcset < Minitest::Test
   include PictureTag
